Test adding random bonus/penalty for hitters
This change is a test of what a bonus or penalty would do to hitter
performance. The methodology isn't sound (we're basically overweighting
singles when things are good and underweighting singles when things are
bad, with no change on other non-out outcomes), but if this has the desired
effect we'll look for a bonus that gets applied correctly across (maybe
having outcomes determined by distance from 0.5 rather than just
shifting the 0.0-1.0 spectrum around).